feat(instance-create) split device section into storage and network
